refactor(ui,app): rename "Finder" to "File Manager" for cross-platform terminology
Houston ships on macOS, Windows, and Linux. "Finder" is macOS-specific,
so user-facing labels and generic shorthand now use a cross-platform
term: "File Manager" in English; "Explorador de Archivos" / "Explorador
de Arquivos" (matching the native Windows naming for File Explorer in
Spanish and Brazilian Portuguese).

User-facing labels (en/es/pt):
- "Open in Finder" -> "Open in File Manager" / "Abrir en el Explorador
  de Archivos" / "Abrir no Explorador de Arquivos"
  (i18n key files.openInFinder -> files.openInFileManager)
- "Show in Finder" -> "Show in File Manager" / "Mostrar en el Explorador
  de Archivos" / "Mostrar no Explorador de Arquivos"
  (file menu reveal label)

Internal symbols (ui/agent):
- finder-icons.tsx -> file-manager-icons.tsx (file rename, git mv)
- formatFinderDate -> formatFileManagerDate (function + all imports)

Cross-platform doc-comments generalized:
- CLAUDE.md / engine.ts: "reveal-in-finder" -> "reveal-in-file-manager"
- os.rs / os-bridge.ts / files.rs / always-on README: "Finder" used as
  generic shorthand replaced with "file manager" / "OS file manager"

Kept as-is (factual references that should stay accurate):
- "macOS Finder-style" / "modeled on Finder" / "format like Finder" in
  ui/agent (icons + utils were specifically modeled on Finder behavior)
- "Finder / Explorer" parentheticals describing real per-platform apps
- houston-composio comments about real macOS .app launch behavior
- "Spotlight, Dock, or Finder" in cli-bundling.md (real macOS launchers)
- use-session-events.ts comment about Finder-initiated user delete

Verified: pnpm typecheck (15 ws projects), pnpm check-locales (es/pt in
sync), cargo check on houston-engine-core, houston-engine-server, and
app/src-tauri.
