+++ ui/agent/src/file-manager-icons.tsx
